Remove InputRegistrationHandler, absorb into Arena
